+++ x/marker/types/codec.go
@@ -30,7 +30,6 @@ func RegisterInterfaces(registry types.InterfaceRegistry) {
 
 	registry.RegisterImplementations(
 		(*govtypesv1beta1.Content)(nil),
-		&AddMarkerProposal{},
 		&SupplyIncreaseProposal{},
 		&SupplyDecreaseProposal{},
 		&SetAdministratorProposal{},