+++ CHANGELOG.md
@@ -58,6 +58,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Add attribute cli command to query account addresses by attribute name [#1451](https://github.com/provenance-io/provenance/issues/1451).
 * Add removal of attributes from accounts on name deletion [#1410](https://github.com/provenance-io/provenance/issues/1410).
 * Enhance ability of smart contracts to use the metadata module [#1280](https://github.com/provenance-io/provenance/issues/1280).
+* Enhance the `AddMarker` endpoint to bypass some validation if issued via governance proposal [#1358](https://github.com/provenance-io/provenance/pull/1358).
+  This replaces the old `AddMarkerProposal` governance proposal.
 
 ### Deprecated
 
@@ -76,10 +78,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Removed the `github.com/provenance-io/provenance/x/metadata/types/p8e` proto package [#1402](https://github.com/provenance-io/provenance/issues/1402).
   Users that generate code from the Provenance protos might need to delete their `p8e/` directory.
 * The `write-scope` CLI command now takes in `[owners]` as semicolon-delimited parties (instead of comma-delimited `[owner-addresses]`) [PR 1453](https://github.com/provenance-io/provenance/pull/1453).
+* Removed the `AddMarkerProposal` [#1358](https://github.com/provenance-io/provenance/pull/1358).
+  It is replaced by putting a `MsgAddMarker` (with the `from_address` of the gov module account), in a `MsgSubmitProposal`.
 
 ### API Breaking
 
 * Removed the `WriteP8eContractSpec` and `P8eMemorializeContract` endpoints [#1402](https://github.com/provenance-io/provenance/issues/1402).
+* Removed the `AddMarkerProposal` [#1358](https://github.com/provenance-io/provenance/pull/1358).
+  It is replaced by putting a `MsgAddMarker` (with the `from_address` of the gov module account), in a `MsgSubmitProposal`.
 
 ### State Machine Breaking